+++ CHANGELOG.md
@@ -4,5 +4,12 @@
 Changelog
 ---------
 
+### v0.5.1
+* Options renamed.
+* Now it gets the set UUID and code from the set name defined at the input file.
+* Do not require ActiveSupport::OrderedHash if Ruby version >= 1.9.
+* Fixed a bug when stripping a card definition string and split it into an array, losing the empty fields.
+* Added unit testings.
+
 ### v0.5.0
 * First release.